Make calling convention Windows 32 bits friendly
x86 calling convention passes arguments through the stack, in opposition
to amd64 where the first arguments go through the CPU registers. This
has the side effect that incompatible calling conventions between C++
and .NET may pass silently in amd64 but cause stack corruption in x86.

This patch fixes the calling convention of exported funtions to be
stdcasll as expected by the TizenFX platform.

Change-Id: Ifbf17d74a769cf011a2451e3d4400b6565537dca